Rename files in the destination package
